refactor: utils 함수 분리 작업
